// src/pageObjects/objects/athena/common/insurance-type-selection/aliases.generated.ts
// src/pageObjects/objects/athena/common/insurance-type-selection/aliases.generated.ts
// AUTO-GENERATED by src/scanner/elements-generator
// pageKey: athena.common.insurance-type-selection
// scannedAt: 2026-03-13T10:57:20.737Z

import type { ElementKey } from "./elements";

export const pageMeta = {
  pageKey: "athena.common.insurance-type-selection",
  urlPath: "/products",
  urlRe: /^\/products$/i,
  title: "Get a new quote or view saved quotes",
  titleRe: new RegExp("Get a new quote or view saved quotes", "i"),
} as const;

export const aliasesGenerated = {
  buttonInputBack: "buttonInputBack" as ElementKey,
  linkCarQuote: "linkCarQuote" as ElementKey,
  linkHomeQuote: "linkHomeQuote" as ElementKey,
  linkInputBack: "linkInputBack" as ElementKey,
  linkLetUsKnow: "linkLetUsKnow" as ElementKey,
  linkTermsAndConditionsApply: "linkTermsAndConditionsApply" as ElementKey,
  linkToAllianzHomePage: "linkToAllianzHomePage" as ElementKey,
} as const;

export type AliasGeneratedKey = keyof typeof aliasesGenerated;